MAINT: Fix Azure

+++ randomgen/distributions.pxd
@@ -1,8 +1,19 @@
 #cython: language_level=3
 
-from libc.stdint cimport (uint8_t, uint16_t, uint32_t, uint64_t,
-                          int8_t, int16_t, int32_t, int64_t, intptr_t)
+from libc.stdint cimport (
+    int8_t,
+    int16_t,
+    int32_t,
+    int64_t,
+    intptr_t,
+    uint8_t,
+    uint16_t,
+    uint32_t,
+    uint64_t,
+)
+
 import numpy as np
+
 cimport numpy as np
 
 ctypedef uint64_t (*next_uint64_t)(void *st) nogil
